feat(back): authorization added

+++ src/controllers/feed.js
@@ -4,6 +4,7 @@ const path = require('path');
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -44,20 +45,31 @@ exports.createPosts = (req, res, next) => {
   }
   const imageUrl = req.file.path;
   const { title, content } = req.body;
+  let creator;
   const post = new Post({
     title,
     content,
     imageUrl,
-    creator: {
-      name: 'Mykola',
-    },
+    creator: req.userId,
   });
   post
     .save()
+    .then(() => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
     .then((result) => {
       res.status(201).json({
         message: 'Post created successfully',
-        post: result,
+        post,
+        creator: {
+          _id: creator._id,
+          name: creator.name,
+        },
       });
     })
     .catch((err) => {
@@ -112,6 +124,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized');
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) clearImage(post.imageUrl);
       post.title = title;
       post.content = content;
@@ -138,12 +155,23 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
-      //?TODO Check if post belongs to logged user
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized');
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then((result) => {
-      res.status(200).json({ post: result, message: 'Post deleted' });
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(() => {
+      res.status(200).json({ message: 'Post deleted' });
     })
     .catch((err) => {
       if (!err.statusCode) {
